+++ include/reactphysics3d/engine/OverlappingPairs.h
@@ -235,7 +235,7 @@ class OverlappingPairs {
                     if (it == lastFrameCollisionInfos.end()) {
 
                         // Make sure capacity is an integral multiple of alignment
-                        const size_t allocatedMemory = std::ceil(sizeof(LastFrameCollisionInfo) / float(GLOBAL_ALIGNMENT)) * GLOBAL_ALIGNMENT;
+                        const size_t allocatedMemory = (size_t)std::ceil(sizeof(LastFrameCollisionInfo) / float(GLOBAL_ALIGNMENT)) * GLOBAL_ALIGNMENT;
 
                         LastFrameCollisionInfo* lastFrameInfo = new (mPoolAllocator->allocate(allocatedMemory)) LastFrameCollisionInfo();
 
@@ -266,7 +266,7 @@ class OverlappingPairs {
                             it->second->LastFrameCollisionInfo::~LastFrameCollisionInfo();
 
                             // Make sure capacity is an integral multiple of alignment
-                            const size_t allocatedMemory = std::ceil(sizeof(LastFrameCollisionInfo) / float(GLOBAL_ALIGNMENT)) * GLOBAL_ALIGNMENT;
+                            const size_t allocatedMemory = (size_t)std::ceil(sizeof(LastFrameCollisionInfo) / float(GLOBAL_ALIGNMENT)) * GLOBAL_ALIGNMENT;
 
                             // Release memory
                             mPoolAllocator->release(it->second, allocatedMemory);
